backend/requirements.txt
Django==6.0.3
dj-database-url==3.1.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
pillow==12.1.1
psycopg[binary]==3.3.3
python-dotenv==1.2.2
reportlab==4.4.10
whitenoise==6.12.0

